requirement.txt
anyio==3.5.0
asgiref==3.5.1
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
dialogflow==2.0.0
fastapi==0.76.0
Flask==2.1.2
google-api-core==1.31.5
google-auth==1.35.0
google-cloud-dialogflow==2.13.0
googleapis-common-protos==1.56.0
grpcio==1.46.0
h11==0.13.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
proto-plus==1.20.3
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.0
pyparsing==3.0.8
pytz==2022.1
requests==2.27.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.18.0
typing-extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
Werkzeug==2.1.2
zipp==3.8.0
